+++ packages/x-components/src/x-modules/queries-preview/store/actions/fetch-and-save-query-preview-results-enrichment.action.ts
@@ -1,0 +1,21 @@
+import type { QueriesPreviewXStoreModule } from '../types'
+import { DefaultResultsEnrichmentService } from '../../../../services/results-enrichment.service'
+
+export const fetchAndSaveQueryPreviewResultsEnrichment: QueriesPreviewXStoreModule['actions']['fetchAndSaveQueryPreviewResultsEnrichment'] =
+  async ({ commit }, queryPreviewItem) => {
+    if (!queryPreviewItem || !queryPreviewItem.results) {
+      return
+    }
+
+    DefaultResultsEnrichmentService.instance
+      .fetchResults(queryPreviewItem.results)
+      .then(enrichmentResults => {
+        commit('updateQueryPreviewResultsFromEnrichment', {
+          queryPreviewItem,
+          enrichmentResults,
+        })
+      })
+      .catch(error => {
+        console.error(error)
+      })
+  }