+++ packages/x-components/src/x-modules/queries-preview/wiring.ts
@@ -72,6 +72,15 @@ export const addQueryPreviewInstanceWire = wireCommit('addQueryPreviewInstance')
 export const removeQueryPreviewInstanceWire = wireCommit('removeQueryPreviewInstance')
 
 /**
+ * Requests and stores the enrichment results for the query preview results.
+ *
+ * @public
+ */
+export const fetchAndSaveQueryPreviewResultsEnrichmentWire = wireDispatch(
+  'fetchAndSaveQueryPreviewResultsEnrichment',
+)
+
+/**
  * Wiring configuration for the {@link QueriesPreviewXModule | queriesPreview module}.
  *
  * @internal
@@ -106,5 +115,8 @@ export const queriesPreviewWiring = createWiring({
   },
   QueryPreviewUnmounted: {
     removeQueryPreviewInstanceWire,
+  },
+  QueryPreviewResultReceived: {
+    fetchAndSaveQueryPreviewResultsEnrichmentWire,
   },
 })